Some changes to default config
Changed MobDropReduction from 0 percent to 100 percent. I believe most users of the plugin would turn down or disable mob drop. otherwise players can easily farm drops near/around spawners. Especially while standing on ladders or standing in hallways; bottle-necking the monsters.

Changed default disabled commands. Expanded with more warp, home, and towny commands that I'm aware of, that would prevent creation of teleport points, other means to teleport in/out, and land claim. Did not add in Worldguard commands as this would usually be a server staff use.

To help not overwhelm players and admins, changed RecallButton to automatically spawn at the end of the dungeon to teleport all in-labyrinth players to the beginning of the labyrinth. Originally disabled.

Changed Over/Under fills to help make sure the traps are generated correctly. As most users will not be looking at the cleanliness of the rooms, as most would be generated underground. With some users using "Air" as one of the block variables, this helps assure the traps are wired and function properly.

Expand Floor/Room/Roof depths to thicken the layers between the floors, again, to make sure traps work and wired. Rooms were heightened to allow the jump damage mechanics come into use.

Increased AirWebPct to 10 percent, to help increase the challenge of navigating the rooms. 1 percent only gave a little bit of an atmosphere.

Added onto the list of natural blocks to include those in caves, dungeons, and mineshafts. 

Need to be added:
 - In the natural blocks, add in hardened clay and colored sand.
 - In the Loot, need to add in new potions, but no all, and other new items of interest. As far back as MC 1.6